add tool processor ext
