+++ OpenAI/Packages/com.openai.unity/Runtime/Threads/CreateThreadAndRunRequest.cs
@@ -4,7 +4,6 @@ using Newtonsoft.Json;
 using OpenAI.Extensions;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using UnityEngine.Scripting;
 
 namespace OpenAI.Threads
@@ -147,42 +146,9 @@ namespace OpenAI.Threads
             AssistantId = assistantId;
             Model = model;
             Instructions = instructions;
-
-            var toolList = tools?.ToList();
-
-            if (toolList is { Count: > 0 })
-            {
-                if (string.IsNullOrWhiteSpace(toolChoice))
-                {
-                    ToolChoice = "auto";
-                }
-                else
-                {
-                    if (!toolChoice.Equals("none") &&
-                        !toolChoice.Equals("required") &&
-                        !toolChoice.Equals("auto"))
-                    {
-                        var tool = toolList.FirstOrDefault(t => t.Function.Name.Contains(toolChoice)) ??
-                                   throw new ArgumentException($"The specified tool choice '{toolChoice}' was not found in the list of tools");
-                        ToolChoice = new { type = "function", function = new { name = tool.Function.Name } };
-                    }
-                    else
-                    {
-                        ToolChoice = toolChoice;
-                    }
-                }
-
-                foreach (var tool in toolList)
-                {
-                    if (tool?.Function?.Arguments != null)
-                    {
-                        // just in case clear any lingering func args.
-                        tool.Function.Arguments = null;
-                    }
-                }
-            }
-
-            Tools = toolList?.ToList();
+            tools.ProcessTools(toolChoice, out var toolList, out var activeTool);
+            Tools = toolList;
+            ToolChoice = activeTool;
             ToolResources = toolResources;
             Metadata = metadata;
             Temperature = reasoningEffort > 0 ? null : temperature;
