+++ OpenAI/Packages/com.openai.unity/Tests/TestFixture_07_Audio.cs
@@ -2,7 +2,6 @@
 
 using NUnit.Framework;
 using OpenAI.Audio;
-using OpenAI.Chat;
 using OpenAI.Models;
 using System.Collections.Concurrent;
 using System.IO;
